// services/query-gen-svc/src/types.ts
export type ConfigMetaType = {
    configId: string;
    configVersion: string;
    configStatus: string;
    configName: string;
    dependentConfig: {
        configId: string;
        configVersion: string;
    };
    creator: string;
    created: string;
    modifier: string;
    modified: string;
};
// ...
export type StudyMriConfigMetaDataType = {
    config: any;
    meta: ConfigMetaType;
    schemaName: string;
};